requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
Pillow==10.1.0
PyPDF2==3.0.1
img2pdf==0.5.0
PyMuPDF==1.23.8
pydantic-settings==2.1.0
aiofiles==23.2.1
